Re-inject partial short options
Given an short option as input, parse the input, resulting in a
`Value` for each matched `Option` from the specification:

    (Argument, Value)

However, because of the ambiguity of short options, we modify the
parser by re-injecting the unparsed rest (the remainder) of the
short option back onto the input stream. This allows to repeatedly
parse pieces of the input until it is fully consumed or we fail.
This, the return value must be:

    (Maybe Token, Maybe (Argument, Value))