Do not parse values w/ commas into lists
Ensure comma separated values provided on the command line or through
the environment are *NOT* parsed into arrays. This has been identified
as flaw in design an is subject to removal from neodoc. This change
excludes default value declarations ([default: a,b,c]) since it's in the
developers control and comma-parsing can be skipped by quoting
([default: 'a,b,c']). This changeset further brings in a fix for the
Value module's read function.

Refer #16