Fix up git tests
The parser has gotten "more correct", which highlighted an issue with
the git-branch tests, in that really the specification as it stands is
ambiguous.

The problem is, how do you decide between:

```
usage: prog a [b]
   or: prog a ...
```

I think part of the solution will be to only substitute [options] for
the 'Generic options' outline in the git-branch help text.