Parse option-arguments repeatedly
Parse option-arguments repeatedly if the actual option-argument is said
to repeat (via `...`). This way multiple arguments can be passed right
after the option:

```
usage: prog --foo=ARG...
```

This would be able to match `$prog --foo 1 2 3` and yield: `{"--foo": [
1, 2, 3 ]}`. The problem, however, is detecting when to stop parsing
since it will eagerly consume any positionals and arguments required for
a successful parse.

The way it's implemented in this commit, the user would have the option
to opt out of this behaviour by wrapping the option in a singleton
group:

```
usage: prog (--foo=ARG)...
```

This way, in order to repeat the option, the user would have to bring
enter multiple `--foo`: `$prog --foo 1 --foo 2 --foo 3`.

Note, however that even with the former usage, the latter is possble:
`$prog --foo 1 2 --foo 3` would be equivalent to `$prog --foo 1 2 3`