Improve foreign-spec failure output
Also update one of the tests to not use '10.0', but use '10.1' instead
in order to preserve the type of the value after serialization. Since
there are no integers in JS, JS sees '10.0' as '10' and the precision
is lost when leaving and re-entering JS otherwise. Supporting this case
is not yet a necessity, so altering a single test is an OK trade off to
make.