Express option aliases as a non-empty list
Leverage the type system to enforce the presence of at least one option
alias. This does a) future proof for unforeseen situations and b) opens
the door to having more than a fixed amount of aliases per option.

Note that these changes currently only have an impact on the Argument
type used for parsing argv (as contained in the specification). Some
conversion between the representation in the `SpecParser.Desc` module
are necessary that highlight - in particular - the weakness of the
current typings, requiring `unsafePartial` calls to compile.