Fix wrong depth calculation
Use a `withLocalDepth` combinator to restore the current depth after
the inner parser ran. This problem became known after implementing
canonicalisation:

```
usage: prog [-a FOO | [[-a FOO] [-b FOO]]]
options:
  -a FOO [default: 100]
  -b FOO [default: 200]
```

Here, `$ a -x -b y` would fail because the first branch was falsly
elected as the winning branch and there is no mention of '-b' on that
branch a parser error occured.