Refactor Neodoc
Hard-to-solve issues that shouldn't be are collecting in the Github
issue tracker. This is a big effort push to clean up many of the core
data structures and implementations.

Most prominently is the move to a more feasible data structure for
describing arguments, using a tree-like ADT:

    data Layout a
        = Group Optional Repeatable (NonEmpty List a)
        | Elem a

This layout is polymorphic to allow for easy transformations on `a`,
e.g. during solving and applying expansion and reduction transforms.

Furthermore the arg parser is being re-written using a custom parser
monad implementation. The previous implementation had significantly
hacked around the limitations of the purescript-parsing parser, such as
the lack of custom errors.

Also worth mentioning:

* Lists that are expected to be non-empty are now encoded as `NonEmpty
  List a` at the type level
* The (pre-)solver spec is expressed entirely in strings, doing away
  with the many helper functions to construct arguments.