Re-implement arg parser
Re-implement the arg parser component of neodoc with performance,
flexibility and maintainability in mind.

The current implementation performs poorly under certain inputs and
configuration. Further, it is under-documented, complex and full of
special cases and one-off fix ups.

This re-implementation addresses these issues from the ground up,
documenting the code along the way in an effort to make it accessible
to outsiders.