Handle implicit flags at parser level
Use a combination of tricks to improve the reliability of implicitly
added flags.

In particular, implicitly added options previously would allow a empty
argv for any program, since flags are always optional. To overcome this,
we make the implicitly added flags non-optional if and only if there is
no other top-level branch that could match the empty input. This is so
that the programmer only has to worry about the empty input if his
explicitly written help text allows the empty input.