Handle implicit arguments at parse level
This is take 3 at implicit argument handling. It turned out to have
more quirks than expected but this final solution appears to cover all
cases. The approach taken now is that we first attempt to parse the spec
"as-is", then re-run a separate parse just for implicit args, e.g.:
"(-h | --help | --version)...", where we make all flags required in
order to trigger a parse fail and not substitute. If this implicit-args
parser fails, we re-throw the previously caught error error.